Adicionar funcionário

+++ MyStock/src/ModeloClasse/Cliente.java
@@ -2,7 +2,7 @@ package ModeloClasse;
 
 public class Cliente extends Usuario {
 
-    public Cliente(String nome, String cpf, String email, int telefone, String uf, String cidade, String endereco, int numero, String bairro) {
+    public Cliente(String nome, String cpf, String email, Long telefone, String uf, String cidade, String endereco, Long numero, String bairro) {
         super(nome, cpf, email, telefone, uf, cidade, endereco, numero, bairro);
     }
